refactor: move ayahInfo extraction to its own service
